options bar component added & photo delete working
